Sudoku Solver Update
I am pleased to announce that the solver can now solve both hard and
easy problems. However, it still has some difficulties with some of the
hard problems. For some reason it won't solve all of them, which is
weird. I'm thinking it may have something to do with the
SinglesUniques() method in the Board class. I will have a look at it
shortly.